requirements.txt
requests
numpy
attrs
scipy
networkx

# For discord bot
discord.py
python-dotenv